Change app directory structure
