+++ apps/batchbook/lib/batchbook.rb
@@ -1,0 +1,2 @@
+require_relative 'batchbook/event_handler.rb'
+require_relative 'batchbook/base.rb'
